More double to float conversion fixes

+++ src/daysim/numerical.h
@@ -1,9 +1,9 @@
 #pragma once
 
-float gammln ( float xx );
-float betai (float a, float b, float x );
-float betacf (float a, float b, float x );
-float gasdev ( long *idum );
+double gammln(double xx);
+double betai(double a, double b, double x);
+double betacf(double a, double b, double x);
+double gasdev(long *idum);
 float ran1 ( long *idum );
 void indexx ( unsigned long n, float *arrin, int *indx );
 void rank ( unsigned long n, int *indx, int *irank );
